# backend/src/team/dto/team-list-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class TeamListResponseDto {
  @ApiProperty()
  id: string;

  @ApiProperty()
  name: string;

  @ApiProperty({ type: String, format: 'date-time' })
  createdAt: Date;

  @ApiProperty({ type: String, format: 'date-time' })
  updatedAt: Date;
}